requirements.txt
boto3
click
dnspython
requests
